requirements.txt
numpy
scipy
matplotlib
gpxpy
requests
pyyaml